requirements.txt
crewai
chromadb
anthropic
openai
python-dotenv
pyyaml
requests
beautifulsoup4
streamlit
pandas
pydantic-settings
voyageai
pymupdf4llm
pdfplumber
langchain
langchain-text-splitters
# Google Drive sync (Phase 9)
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pytrends>=4.9
